refactor(vite): extract license loss entity hook
