+++ vite/src/components/forms/attach-v2/index.ts
@@ -22,6 +22,7 @@ export * from "./hooks/useAttachMutation";
 export * from "./hooks/useAttachPreview";
 export * from "./hooks/useAttachRequestBody";
 export * from "./hooks/useGrantFree";
+export * from "./hooks/useLicenseLossEntities";
 export * from "./hooks/usePreviewDiff";
 
 // Utils
